Fix float64/float32 dtype mismatch in quantum model outputs
PennyLane's default.qubit device returns float64 expectation values
regardless of the torch input dtype. On the installed PennyLane version
(0.45.1, newer than the >=0.38.0 floor in requirements.txt) this broke
QCRBM, QFeatureQRBM, and QQRBM with "expected m1 and m2 to have the same
dtype, but got: double != float" as soon as the quantum features were
multiplied against the float32 classical parameters. Cast each qnode's
stacked output back to float32 at the source.

Verified with check_environment.py: all 6 checks now pass (previously
3/6 failed on this exact error).

+++ src/models/qcrbm.py
@@ -117,7 +117,9 @@ class QCRBM(CRBM):
         broadcasts over a batch dimension when the input tensor has one."""
         x = self._encode_input(v, u)
         out = self.circuit(x, self.theta)  # list of H tensors, each (batch,)
-        return torch.stack(out, dim=1)  # (batch, H)
+        # PennyLane's default.qubit device returns float64 regardless of the
+        # torch input dtype; cast back to float32 to match the rest of the model.
+        return torch.stack(out, dim=1).to(torch.float32)  # (batch, H)
 
     # -- augmented hidden conditional (Eq. 22) --------------------------
     def p_h_given_v_quantum(self, v: torch.Tensor, u: torch.Tensor, q: torch.Tensor = None):
